Split gui and msgbox into separate modules
